refactor(RELEASE-1009): remove upload_sbom.py pyxis script
This commit removes pyxis/upload_sbom.py and its test file as it is no
longer used in our pipelines. Some of its functions were used in
upload_rpm_data.py, so those and their tests were migrated to the
upload_rpm_data.py file.
