--- pyxis/upload_sbom
@@ -1,1 +1,0 @@
-upload_sbom.py
